refactor: Customer의 멤버를 VO로 수정
