+++ src/main/java/woowacourse/shoppingcart/domain/user/Nickname.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.domain;
+package woowacourse.shoppingcart.domain.user;
 
 import woowacourse.shoppingcart.exception.InvalidNicknameException;
 
@@ -8,8 +8,16 @@ public class Nickname {
 
     public Nickname(String nickname) {
         this.nickname = nickname;
-        if (!this.nickname.matches("^[가-힣A-Za-z0-9]{2,8}$")) {
+        validate();
+    }
+
+    private void validate() {
+        if (!nickname.matches("^[가-힣A-Za-z0-9]{2,8}$")) {
             throw new InvalidNicknameException();
         }
+    }
+
+    public String value() {
+        return nickname;
     }
 }
