+++ src/test/java/woowacourse/shoppingcart/domain/CustomerTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import woowacourse.shoppingcart.domain.user.Customer;
+import woowacourse.shoppingcart.domain.user.Password;
 
 public class CustomerTest {
 
@@ -14,6 +16,6 @@ public class CustomerTest {
     void isValidPassword(String password, boolean expected) {
         Customer customer = new Customer("email@email.com", "12345678a", "tonic");
 
-        assertThat(customer.isValidPassword(password)).isEqualTo(expected);
+        assertThat(customer.isValidPassword(new Password(password))).isEqualTo(expected);
     }
 }
