+++ src/test/java/woowacourse/shoppingcart/domain/NicknameTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import woowacourse.shoppingcart.domain.user.Nickname;
 import woowacourse.shoppingcart.exception.InvalidNicknameException;
 
 public class NicknameTest {
